requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# AI 相关
openai>=1.50.0
anthropic==0.7.7
requests==2.31.0

# PPT 处理
python-pptx==0.6.23
Pillow==10.1.0
python-docx==1.1.0

# 数据处理
pandas==2.1.3
numpy>=1.24.0
markdown==3.5.1
beautifulsoup4==4.12.2

# 数据库
sqlalchemy==2.0.23

# 工具库
pydantic>=2.8.0
pydantic-settings>=2.6.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 开发工具
python-dotenv==1.0.0

# PDF 处理
PyPDF2==3.0.1 